test: split post-process coverage by script
