+++ tests/simulations/post_process/test_produce_cmatrix_table.py
@@ -1,0 +1,30 @@
+# This code is part of KQCircuits
+# Copyright (C) 2026 IQM Finland Oy
+#
+# This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
+# License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later
+# version.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with this program. If not, see
+# https://www.gnu.org/licenses/gpl-3.0.html.
+#
+# The software distribution should follow IQM trademark policy for open-source software
+# (meetiqm.com/iqm-open-source-trademark-policy). IQM welcomes contributions to the code.
+# Please see our contribution agreements for individuals (meetiqm.com/iqm-individual-contributor-license-agreement)
+# and organizations (meetiqm.com/iqm-organization-contributor-license-agreement).
+
+from post_process_test_helpers import read_csv, run_post_process, write_json
+
+
+def test_produce_cmatrix_table_writes_capacitance_matrix_csv(tmp_path, monkeypatch):
+    """`produce_cmatrix_table.py` turns mocked project capacitances into a result table."""
+    write_json(tmp_path / "waveguide.json", {"parameters": {}, "tool": "cross-section"})
+    write_json(tmp_path / "waveguide_project_results.json", {"CMatrix": [[1.0, 2.0], [3.0, 4.0]]})
+
+    run_post_process("produce_cmatrix_table.py", tmp_path, monkeypatch, __file__)
+
+    rows = read_csv(tmp_path / f"{tmp_path.name}_results.csv")
+    assert rows == [{"key": "waveguide", "C11": "1.0", "C12": "2.0", "C21": "3.0", "C22": "4.0"}]
